Moved RandomBytes function from utils in vm package to utils in protocol package since it is used by other components
